moved fiber error handler to fiber package
