+++ internal/handler/decision/module.go
@@ -15,10 +15,10 @@ import (
 
 	"github.com/dadrus/heimdall/internal/cache"
 	"github.com/dadrus/heimdall/internal/config"
+	"github.com/dadrus/heimdall/internal/fiber/errorhandler"
 	fibercache "github.com/dadrus/heimdall/internal/fiber/middleware/cache"
 	fiberlogger "github.com/dadrus/heimdall/internal/fiber/middleware/logger"
 	fibertracing "github.com/dadrus/heimdall/internal/fiber/middleware/tracing"
-	"github.com/dadrus/heimdall/internal/handler/errorhandler"
 	"github.com/dadrus/heimdall/internal/x"
 )
